+++ scripts/test/Console/Command/Admin/Config/FixInvalid.t
@@ -12,7 +12,6 @@ use utf8;
 
 use vars (qw($Self));
 
-# get helper object
 $Kernel::OM->ObjectParamAdd(
     'Kernel::System::UnitTest::Helper' => {
         RestoreDatabase => 1,
@@ -102,7 +101,8 @@ $Self->Is(
 
 $Self->True(
     (
-        $Result =~ m{Auto-corrected setting: Ticket::Frontend::AgentTicketPhone###Priority.*?Deployment successful\.}s
+        $Result
+            =~ m{Auto-corrected setting:.*Ticket::Frontend::AgentTicketPhone###Priority.*Deployment successful\.}s
     ) // 0,
     'Check invalid result'
 );